Rename `dst_rule.py` to `transition_rule.py` and update imports for clarity and consistency.
